Added object into an arrat in index.js
