+++ src/index.html
@@ -1,9 +1,7 @@
-<!DOCTYPE html>
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>to do list</title>
     <link
       rel="stylesheet"
@@ -21,12 +19,12 @@
           type="text"
           placeholder="Add to your list..."
         />
-        <button type="button" class="add-button">
+        <button type="button" class="add-buttom">
           <i class="fas fa-sign-in-alt"></i>
         </button>
       </div>
       <ul class="todo-list"></ul>
-      <button type="button" class="clear-all">Clear the completed tasks</button>
+      <button type="button" class="clear-all">Clear all completed</button>
     </div>
   </body>
 </html>
