+++ src/style.css
@@ -1,29 +1,8 @@
-* {
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-}
-
-body {
-  background-color: #fff;
-}
-
 .container {
   display: flex;
   flex-direction: column;
   align-items: center;
-  width: 80%;
-  margin-top: 8rem;
-  border: 2px solid #060607c7;
-  margin-left: 10%;
-  box-sizing: border-box;
-  padding: 9px;
-  border-radius: 10px;
-}
-
-.container h2 {
-  align-self: center;
-  margin-bottom: 10px;
+  width: 100%;
 }
 
 #todo-list {
@@ -34,48 +13,24 @@ body {
 ul {
   list-style: none;
   padding: 0;
-  width: 100%;
-  margin: 0;
 }
 
 li {
-  padding: 2% 7px;
-  border-top: 2px solid lightgray;
-  width: 100%;
+  border-bottom: 2px solid lightgray;
+  padding: 6%;
+  width: 240px;
   display: flex;
   justify-content: space-between;
 }
 
 .input-container {
   display: flex;
-  width: 98%;
+  width: 100%;
   justify-content: center;
 }
 
-input {
-  border: 1px solid transparent;
-}
-
 .add-input {
-  width: 100%;
-  height: 30px;
-  padding: 5px;
-}
-
-.add-button {
-  background: transparent;
-  border: transparent;
-  margin-left: 5px;
-}
-
-.clear-all {
-  width: 102%;
-  height: 43px;
-  border-radius: 0 0 42px 40px;
-  padding: 5px;
-  border: transparent;
-  background: #ededed59;
-  cursor: pointer;
+  width: 45%;
 }
 
 .task {
@@ -84,24 +39,7 @@ input {
   border-bottom: 1px solid #ccc;
 }
 
-.todo-completed .todo-item-edit {
+.todo-completed {
   text-decoration: line-through;
   color: #ccc;
-}
-
-.icon {
-  width: 17px;
-  background: url(./sUqME.jpg);
-  background-position: center;
-  background-size: contain;
-  background-repeat: no-repeat;
-  cursor: pointer;
-}
-
-.icon:hover {
-  background: url(./1175088.png);
-  width: 15px;
-  background-position: center;
-  background-repeat: no-repeat;
-  background-size: contain;
 }
